backend/requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
matplotlib
seaborn
fastapi
uvicorn
python-dotenv
joblib
pytest
streamlit==1.19.0
altair==4.2.2
-e .